requirements.txt
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
factory-boy==2.12.0
